+++ OsvvmProjectScripts.tcl
@@ -19,12 +19,20 @@
 # 
 #  Revision History:
 #    Date      Version    Description
-#    11/2018   Alpha      Project descriptors in .files and .dirs files
+#     2/2021   2021.02    Updated initialization of libraries                 
+#                         Analyze allows ".vhdl" extensions as well as ".vhd" 
+#                         Include/Build signal error if nothing to run                         
+#                         Added SetVHDLVersion / GetVHDLVersion to support 2019 work           
+#                         Added SetSimulatorResolution / GetSimulatorResolution to support GHDL
+#                         Added beta of LinkLibrary to support linking in project libraries    
+#                         Added beta of SetLibraryDirectory / GetLibraryDirectory              
+#                         Added beta of ResetRunLibrary                                        
+#     7/2020   2020.07    Refactored tool execution for simpler vendor customization
+#     1/2020   2020.01    Updated Licenses to Apache
 #     2/2019   Beta       Project descriptors in .pro which execute 
 #                         as TCL scripts in conjunction with the library 
 #                         procedures
-#     1/2020   2020.01    Updated Licenses to Apache
-#     7/2020   2020.07    Refactored tool execution for simpler vendor customization
+#    11/2018   Alpha      Project descriptors in .files and .dirs files
 #
 #
 #  This file is part of OSVVM.
@@ -48,7 +56,7 @@
 # StartUp
 #   re-run the startup scripts, this program included
 #
-proc StartUp {} {
+proc StartUp {} { 
   puts "source $::SCRIPT_DIR/StartUp.tcl"
   source $::SCRIPT_DIR/StartUp.tcl
 }
@@ -116,6 +124,14 @@ proc StopTranscript {{FileBaseName ""}} {
   }
 }
 
+proc TerminateTranscript {} {
+  global OsvvmCurrentTranscript
+
+  if {[info exists OsvvmCurrentTranscript]} {
+    set OsvvmCurrentTranscript ""
+  }
+}
+
 #
 #  Problematic since output of tests has the word log
 #
@@ -129,8 +145,13 @@ proc StopTranscript {{FileBaseName ""}} {
 #
 proc include {Path_Or_File} {
   global CURRENT_WORKING_DIRECTORY
+  global VHDL_WORKING_LIBRARY
   
 #  puts "set StartingPath ${CURRENT_WORKING_DIRECTORY} Starting Include"
+  # If a library does not exist, then create the default
+  if {![info exists VHDL_WORKING_LIBRARY]} {
+    library default
+  }
   set StartingPath ${CURRENT_WORKING_DIRECTORY}
   
   set NormName [file normalize ${StartingPath}/${Path_Or_File}]
@@ -172,27 +193,36 @@ proc include {Path_Or_File} {
     set FileTclName    ${FileBaseName}.tcl
     set FileDoName     ${FileBaseName}.do
 
+    set FoundActivity 0
     if {[file exists ${FileProName}]} {
       puts "source ${FileProName}"
       source ${FileProName} 
+      set FoundActivity 1
     } 
     # .dirs is intended to be deprecated in favor of .pro
     if {[file exists ${FileDirsName}]} {
       Do_List ${FileDirsName} "include"
+      set FoundActivity 1
     }
     # .files is intended to be deprecated in favor of .pro
     if {[file exists ${FileFilesName}]} {
       Do_List ${FileFilesName} "analyze"
+      set FoundActivity 1
     }
     # .tcl intended for extended capability
     if {[file exists ${FileTclName}]} {
       puts "do ${FileTclName} ${CURRENT_WORKING_DIRECTORY}"
       eval do ${FileTclName} ${CURRENT_WORKING_DIRECTORY}
+      set FoundActivity 1
     }
     # .do intended for extended capability
     if {[file exists ${FileDoName}]} {
       puts "do ${FileDoName} ${CURRENT_WORKING_DIRECTORY}"
       eval do ${FileDoName} ${CURRENT_WORKING_DIRECTORY}
+      set FoundActivity 1
+    }
+    if {$FoundActivity == 0} {
+      error "Build / Include did not find anything to execute"
     }
   } 
 #  puts "set CURRENT_WORKING_DIRECTORY ${StartingPath} Ending Include"
@@ -201,12 +231,8 @@ proc include {Path_Or_File} {
 
 proc build {{Path_Or_File "."} {LogName "."}} {
   global CURRENT_WORKING_DIRECTORY
-  global LIB_BASE_DIR
-  global OSVVM_SCRIPTS_INITIALIZED
-  global ToolNameVersion 
-  global DIR_LIB
-  global DIR_LOGS
   global CURRENT_RUN_DIRECTORY
+  global VHDL_WORKING_LIBRARY
 
   set CURRENT_WORKING_DIRECTORY [pwd]
   
@@ -214,28 +240,16 @@ proc build {{Path_Or_File "."} {LogName "."}} {
     set CURRENT_RUN_DIRECTORY ""
   }
 
-  
-  # Create 
-  if {![info exists OSVVM_SCRIPTS_INITIALIZED] || $CURRENT_WORKING_DIRECTORY ne $CURRENT_RUN_DIRECTORY } {
-    set OSVVM_SCRIPTS_INITIALIZED 1
-    
-    set CURRENT_RUN_DIRECTORY $CURRENT_WORKING_DIRECTORY
-  
-    if {![info exists LIB_BASE_DIR]} {
-      set LIB_BASE_DIR $CURRENT_WORKING_DIRECTORY
+  # Initialize 
+  if {![info exists VHDL_WORKING_LIBRARY] || $CURRENT_WORKING_DIRECTORY ne $CURRENT_RUN_DIRECTORY } {
+    if {[info exists VHDL_WORKING_LIBRARY]} {
+      unset VHDL_WORKING_LIBRARY
     }
-    
-    # Set locations for libraries and logs
-    set DIR_LIB    ${LIB_BASE_DIR}/VHDL_LIBS/${ToolNameVersion}
-    set DIR_LOGS   ${CURRENT_WORKING_DIRECTORY}/logs/${ToolNameVersion}
-
-    # Create LIB and Results directories
-    CreateDirectory $DIR_LIB
-    CreateDirectory ${CURRENT_WORKING_DIRECTORY}/results
-
-    # Create default library
-    library default
-  }
+    library default 
+  } 
+  
+  # If Transcript Open, then Close it
+  TerminateTranscript
   
   # Create the Log File Name
   set NormPathOrFile [file normalize ${Path_Or_File}]
@@ -282,11 +296,45 @@ proc CreateDirectory {Directory} {
 }
 
 # -------------------------------------------------
+# OsvvmInitialize
+#
+proc OsvvmInitialize {} {
+  global CURRENT_WORKING_DIRECTORY
+  global CURRENT_RUN_DIRECTORY
+  global LIB_BASE_DIR
+  global DIR_LIB
+  global DIR_LOGS
+  global ToolNameVersion 
+
+  if {![info exists CURRENT_WORKING_DIRECTORY]} {
+    set CURRENT_WORKING_DIRECTORY [pwd]
+  }
+  set CURRENT_RUN_DIRECTORY [pwd]
+
+  if {![info exists LIB_BASE_DIR]} {
+    set LIB_BASE_DIR $CURRENT_RUN_DIRECTORY
+  }
+  
+  # Set locations for libraries and logs
+  set DIR_LIB    ${LIB_BASE_DIR}/VHDL_LIBS/${ToolNameVersion}
+  set DIR_LOGS   ${CURRENT_RUN_DIRECTORY}/logs/${ToolNameVersion}
+
+  # Create LIB and Results directories
+  CreateDirectory $DIR_LIB
+  CreateDirectory ${CURRENT_RUN_DIRECTORY}/results
+}
+
+# -------------------------------------------------
 # Library
 #
 proc library {LibraryName} {
-  global DIR_LIB
   global VHDL_WORKING_LIBRARY
+  global DIR_LIB
+  
+  # If VHDL_WORKING_LIBRARY does not exist, then initialize
+  if {![info exists VHDL_WORKING_LIBRARY]} {
+    OsvvmInitialize
+  }
   
   puts "library $LibraryName" 
 
@@ -300,6 +348,10 @@ proc library {LibraryName} {
 proc map {LibraryName {PathToLib ""}} {
   global DIR_LIB
   global VHDL_WORKING_LIBRARY
+
+  if {![info exists VHDL_WORKING_LIBRARY]} {
+    OsvvmInitialize
+  }
 
   if {![string match $PathToLib ""]} {
     # only for mapping to external existing library
@@ -319,16 +371,23 @@ proc map {LibraryName {PathToLib ""}} {
 #
 proc analyze {FileName} {
   global CURRENT_WORKING_DIRECTORY
+  global VHDL_WORKING_LIBRARY
+
+  # If a library does not exist, then create the default
+  if {![info exists VHDL_WORKING_LIBRARY]} {
+    library default
+  }
   
   puts "analyze $FileName"
   
   set NormFileName [file normalize ${CURRENT_WORKING_DIRECTORY}/${FileName}]
+  set FileExtension [file extension $FileName]
 
-  if {[file extension $FileName] eq ".vhd"} {
-    vendor_analyze_vhdl $::VHDL_WORKING_LIBRARY ${NormFileName}
-  } elseif {[file extension $FileName] eq ".v"} {
-    vendor_analyze_verilog $::VHDL_WORKING_LIBRARY ${NormFileName}
-  } elseif {[file extension $FileName] eq ".lib"} {
+  if {$FileExtension eq ".vhd" || $FileExtension eq ".vhdl"} {
+    vendor_analyze_vhdl ${VHDL_WORKING_LIBRARY} ${NormFileName}
+  } elseif {$FileExtension eq ".v"} {
+    vendor_analyze_verilog ${VHDL_WORKING_LIBRARY} ${NormFileName}
+  } elseif {$FileExtension eq ".lib"} {
     #  for handling older deprecated file format
     library [file rootname $FileName]
   }
@@ -341,6 +400,10 @@ proc simulate {LibraryUnit {OptionalCommands ""}} {
   global VHDL_WORKING_LIBRARY
 #  StartTranscript ${LibraryUnit}.log
   
+  # If a library does not exist, then create the default
+  if {![info exists VHDL_WORKING_LIBRARY]} {
+    library default
+  }
   set StartTime   [clock seconds] 
   puts "Start time [clock format $StartTime -format %T]"
 
@@ -353,8 +416,137 @@ proc simulate {LibraryUnit {OptionalCommands ""}} {
 #  StopTranscript ${LibraryUnit}.log
 }
 
+
+
+# -------------------------------------------------
+# Settings
+#
+proc SetVHDLVersion {VhdlVersion} {
+  global OsvvmVhdlVersion
+  global OsvvmVhdlShortVersion
+  
+  if {$VhdlVersion eq "2008" || $VhdlVersion eq "08"} {
+    set OsvvmVhdlVersion 2008
+    set OsvvmVhdlShortVersion 08
+  } elseif {$VhdlVersion eq "2019" || $VhdlVersion eq "19" } {
+    set OsvvmVhdlVersion 2019
+    set OsvvmVhdlShortVersion 19
+  } elseif {$VhdlVersion eq "2002" || $VhdlVersion eq "02" } {
+    set OsvvmVhdlVersion 2002
+    set OsvvmVhdlShortVersion 02
+    puts "\nWARNING:  VHDL Version set to 2002.  OSVVM Requires 2008 or newer\n"
+  } elseif {$VhdlVersion eq "1993" || $VhdlVersion eq "93" } {
+    set OsvvmVhdlVersion 93
+    set OsvvmVhdlShortVersion 93
+    puts "\nWARNING:  VHDL Version set to 1993.  OSVVM Requires 2008 or newer\n"
+  } else {
+    set OsvvmVhdlVersion 2008
+    set OsvvmVhdlShortVersion 08
+    puts "\nWARNING:  Input to SetVHDLVersion not recognized.   Using 2008.\n"
+  }  
+}
+
+proc GetVHDLVersion {} {
+  return $::OsvvmVhdlVersion
+}
+
+proc SetSimulatorResolution {SimulatorResolution} {
+  global SIMULATE_TIME_UNITS
+
+  set SIMULATE_TIME_UNITS        $SimulatorResolution
+}
+
+proc GetSimulatorResolution {} {
+  return $::SIMULATE_TIME_UNITS
+}
+
+#
+# Remaining proc are Experimental, Alpha code and are likely to change.
+# Use at your own risk.
+#
+
+#
+#  Currently only set in OsvvmScriptDefaults
+#
+proc SetLibraryDirectory {{LibraryDirectory ""}} {
+  global CURRENT_RUN_DIRECTORY
+  global LIB_BASE_DIR
+  global VHDL_WORKING_LIBRARY
+  global ToolNameVersion
+  
+  if {$LibraryDirectory eq ""} {
+    if {[info exists CURRENT_RUN_DIRECTORY]} {
+      set LIB_BASE_DIR $CURRENT_RUN_DIRECTORY
+      set DIR_LIB    ${LIB_BASE_DIR}/VHDL_LIBS/${ToolNameVersion}
+    } else {
+      # Instead, will be set by first call to build, include, analyze, simulate, or library
+      if {[info exists LIB_BASE_DIR]} {
+        unset LIB_BASE_DIR
+      }
+      if {[info exists VHDL_WORKING_LIBRARY]} {
+        unset VHDL_WORKING_LIBRARY
+      }
+    }
+  } else {
+    set LIB_BASE_DIR $LibraryDirectory
+    set DIR_LIB    ${LIB_BASE_DIR}/VHDL_LIBS/${ToolNameVersion}
+  }
+}
+
+proc GetLibraryDirectory {} {
+  global LIB_BASE_DIR
+  
+  if {[info exists LIB_BASE_DIR]} {
+    return "${LIB_BASE_DIR}"
+  } else {
+    puts "WARNING:  GetLibraryDirectory LIB_BASE_DIR not defined"
+    return ""
+  }
+}
+  
+proc ResetRunDirectory {} {
+  global CURRENT_RUN_DIRECTORY
+  global LIB_BASE_DIR
+  global VHDL_WORKING_LIBRARY
+  
+  if {[info exists CURRENT_RUN_DIRECTORY]} {
+    unset CURRENT_RUN_DIRECTORY
+  }
+  if {[info exists LIB_BASE_DIR]} {
+    unset LIB_BASE_DIR
+  }
+  if {[info exists VHDL_WORKING_LIBRARY]} {
+    unset VHDL_WORKING_LIBRARY
+  }
+}
+
+proc LinkLibrary {{LibraryDirectory ""}} {
+  global CURRENT_RUN_DIRECTORY
+  global ToolNameVersion
+  
+  if {$LibraryDirectory eq ""} {
+    if {[info exists DIR_LIB]} {
+      set CurrentLib $DIR_LIB
+    } else {
+      set CurrentLib ${CURRENT_RUN_DIRECTORY}/VHDL_LIBS/${ToolNameVersion}
+    }
+  } else {
+      set CurrentLib ${LibraryDirectory}/VHDL_LIBS/${ToolNameVersion}
+  }
+  if {[file isdirectory $CurrentLib]} {
+    foreach LibToLink [glob -directory $CurrentLib *] {
+      set LibName [file rootname [file tail $LibToLink]]
+      library $LibName
+    }  
+  } else {
+    puts "$CurrentLib does not exist"
+  }
+}
+
 # -------------------------------------------------
 # MapLibraries
+#   Likely this will be replaced by LinkLibrary.
+#
 #   Used to create a library mapping in a  
 #   directory different from the initial/normal simulation 
 #   directory.