+++ VendorScripts_ActiveHDL.tcl
@@ -19,18 +19,18 @@
 # 
 #  Revision History:
 #    Date      Version    Description
-#    11/2018   Alpha      Project descriptors in .files and .dirs files
+#     2/2021   2021.02    Refactored variable settings to here from ToolConfiguration.tcl
+#     7/2020   2020.07    Refactored tool execution for simpler vendor customization
+#     1/2020   2020.01    Updated Licenses to Apache
 #     2/2019   Beta       Project descriptors in .pro which execute 
 #                         as TCL scripts in conjunction with the library 
 #                         procedures
-#     1/2020   2020.01    Updated Licenses to Apache
-#     7/2020   2020.07    Refactored for simpler vendor customization
-#     7/2020   2020.07    Refactored tool execution for simpler vendor customization
+#    11/2018   Alpha      Project descriptors in .files and .dirs files
 #
 #
 #  This file is part of OSVVM.
 #  
-#  Copyright (c) 2018 - 2020 by SynthWorks Design Inc.  
+#  Copyright (c) 2018 - 2021 by SynthWorks Design Inc.  
 #  
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -44,6 +44,18 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
+
+
+# -------------------------------------------------
+# Tool Settings
+#
+  set ToolType    "simulator"
+  set ToolVendor  "Aldec"
+  set simulator   "ActiveHDL"
+  set ToolNameVersion ${simulator}-${version}
+  puts $ToolNameVersion
+  # Allow global OSVVM library to be updated
+  setlibrarymode -rw osvvm
 
 
 # -------------------------------------------------
@@ -111,6 +123,7 @@ proc vendor_map {LibraryName ResolvedPathToLib} {
 # analyze
 #
 proc vendor_analyze_vhdl {LibraryName FileName} {
+  global OsvvmVhdlVersion
   global DIR_LIB
   
   set MY_START_DIR [pwd]
@@ -120,12 +133,12 @@ proc vendor_analyze_vhdl {LibraryName FileName} {
   if {![file isfile ${DIR_LIB}/$LibraryName/src/${FileBaseName}.vcom]} {
     echo addfile ${FileName}
     addfile ${FileName}
-    filevhdloptions -2008 ${FileName}
+    filevhdloptions -${OsvvmVhdlVersion} ${FileName}
   }
   # Compile it.
-  echo vcom -2008 -dbg -relax -work ${LibraryName} ${FileName} 
-  echo vcom -2008 -dbg -relax -work ${LibraryName} ${FileName} > ${DIR_LIB}/$LibraryName/src/${FileBaseName}.vcom
-  eval vcom -2008 -dbg -relax -work ${LibraryName} ${FileName}
+  echo vcom -${OsvvmVhdlVersion} -dbg -relax -work ${LibraryName} ${FileName} 
+  echo vcom -${OsvvmVhdlVersion} -dbg -relax -work ${LibraryName} ${FileName} > ${DIR_LIB}/$LibraryName/src/${FileBaseName}.vcom
+  eval vcom -${OsvvmVhdlVersion} -dbg -relax -work ${LibraryName} ${FileName}
   
   cd $MY_START_DIR
 }