+++ VendorScripts_GHDL.tcl
@@ -19,12 +19,13 @@
 # 
 #  Revision History:
 #    Date      Version    Description
+#     2/2021   2021.02    Refactored variable settings to here from ToolConfiguration.tcl
 #     9/2020   2020.09    Initial Version
 #
 #
 #  This file is part of OSVVM.
 #  
-#  Copyright (c) 2018 - 2020 by SynthWorks Design Inc.  
+#  Copyright (c) 2020 - 2021 by SynthWorks Design Inc.  
 #  
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -41,23 +42,38 @@
 
 
 # -------------------------------------------------
+# Tool Settings
+#
+  set ToolType   "simulator"
+  set ToolVendor "GHDL"
+  set simulator  "GHDL"
+  set ghdl "ghdl"
+  # required for mintty
+  set console "/dev/pty0"
+  set ToolNameVersion "GHDL-v0.37.0-1063-gc5b094bb-2020-1023"
+  puts $ToolNameVersion
+
+
+# -------------------------------------------------
 # StartTranscript / StopTranscxript
 #
 proc vendor_StartTranscript {FileName} {
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
    
-  if {[info exists GHDL_TRANSCIPT_FILE]} {
-    unset GHDL_TRANSCIPT_FILE 
+  if {[info exists GHDL_TRANSCRIPT_FILE]} {
+    unset GHDL_TRANSCRIPT_FILE 
   }
-  set GHDL_TRANSCIPT_FILE $FileName
-  puts "Transcript $GHDL_TRANSCIPT_FILE" 
+  set GHDL_TRANSCRIPT_FILE $FileName
+  puts "Transcript $GHDL_TRANSCRIPT_FILE" 
+  exec echo "Start Time [clock format [clock seconds] -format %T]" > $GHDL_TRANSCRIPT_FILE
 }
 
 proc vendor_StopTranscript {FileName} {
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
    
-#  unset GHDL_TRANSCIPT_FILE 
-  puts "Stop Transcript $GHDL_TRANSCIPT_FILE" 
+#  unset GHDL_TRANSCRIPT_FILE 
+  puts "Stop Transcript $GHDL_TRANSCRIPT_FILE" 
+  exec echo "Stop Time [clock format [clock seconds] -format %T]" >> $GHDL_TRANSCRIPT_FILE
 }
 
 
@@ -67,7 +83,7 @@ proc vendor_StopTranscript {FileName} {
 proc vendor_library {LibraryName PathToLib} {
   global VHDL_WORKING_LIBRARY_PATH
 #  global VHDL_RESOURCE_LIBRARY_PATHS
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
    
 #  set PathAndLib ${PathToLib}/${LibraryName}.lib
   set PathAndLib ${PathToLib}/${LibraryName}/v08
@@ -92,13 +108,13 @@ proc vendor_library {LibraryName PathToLib} {
 
 proc vendor_map {LibraryName PathToLib} {
   global VHDL_WORKING_LIBRARY_PATH
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
 
   set PathAndLib ${PathToLib}/${LibraryName}.lib
 
   if {![file exists ${PathAndLib}]} {
     error "Map:  Creating library ${PathAndLib} since it does not exist.  "
-    puts "creating library directory ${PathAndLib}" | tee -a $GHDL_TRANSCIPT_FILE
+    puts "creating library directory ${PathAndLib}" 
     file mkdir   ${PathAndLib}
   }
   set VHDL_WORKING_LIBRARY_PATH  ${PathAndLib}
@@ -108,20 +124,22 @@ proc vendor_map {LibraryName PathToLib} {
 # analyze
 #
 proc vendor_analyze_vhdl {LibraryName FileName} {
-  global ghdl console
+  global OsvvmVhdlShortVersion
+  global ghdl 
+  global console
   global VHDL_WORKING_LIBRARY_PATH
 #  global VHDL_RESOURCE_LIBRARY_PATHS
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
   global DIR_LIB
 
 #  puts "$ghdl -a --std=08 -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${FileName}" 
-#  eval exec $ghdl -a --std=08 -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${FileName} | tee -a $GHDL_TRANSCIPT_FILE $console
-  puts "$ghdl -a --std=08 -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${FileName}" 
-  eval exec $ghdl -a --std=08 -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${FileName} | tee -a $GHDL_TRANSCIPT_FILE $console
+#  eval exec $ghdl -a --std=08 -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${FileName} | tee -a $GHDL_TRANSCRIPT_FILE $console
+  exec echo "$ghdl -a --std=${OsvvmVhdlShortVersion} -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${FileName}" | tee -a $GHDL_TRANSCRIPT_FILE $console
+  eval exec $ghdl -a --std=${OsvvmVhdlShortVersion} -Wno-hide --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${FileName} |& tee -a $GHDL_TRANSCRIPT_FILE $console
 }
 
 proc vendor_analyze_verilog {LibraryName FileName} {
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
 
   puts "Analyzing verilog files not supported by GHDL" 
 }
@@ -130,15 +148,16 @@ proc vendor_analyze_verilog {LibraryName FileName} {
 # Simulate
 #
 proc vendor_simulate {LibraryName LibraryUnit OptionalCommands} {
-  global ghdl console
+  global OsvvmVhdlShortVersion
+  global ghdl 
+  global console
   global VHDL_WORKING_LIBRARY_PATH
 #  global VHDL_RESOURCE_LIBRARY_PATHS
-  global GHDL_TRANSCIPT_FILE
+  global GHDL_TRANSCRIPT_FILE
   global DIR_LIB
 
 #  puts "$ghdl --elab-run --std=08 --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${LibraryUnit}" 
-#  eval exec $ghdl --elab-run --std=08 --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${LibraryUnit} | tee -a $GHDL_TRANSCIPT_FILE $console
-  puts "$ghdl --elab-run --std=08 --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${LibraryUnit}" 
-  eval exec $ghdl --elab-run --std=08 --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${LibraryUnit} | tee -a $GHDL_TRANSCIPT_FILE $console
-
+#  eval exec $ghdl --elab-run --std=08 --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} ${VHDL_RESOURCE_LIBRARY_PATHS} ${LibraryUnit} | tee -a $GHDL_TRANSCRIPT_FILE $console
+  exec echo "$ghdl --elab-run --std=${OsvvmVhdlShortVersion} --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${LibraryUnit}" | tee -a $GHDL_TRANSCRIPT_FILE $console
+  eval exec $ghdl --elab-run --std=${OsvvmVhdlShortVersion} --work=${LibraryName} --workdir=${VHDL_WORKING_LIBRARY_PATH} -P${DIR_LIB} ${LibraryUnit} |& tee -a $GHDL_TRANSCRIPT_FILE $console
 }